requirements.txt
tensorflow==2.10.0
numpy
pypinyin==0.49.0
discord.py==2.0.1